--- a/examples/react/large-file-based/src/routes/__root.tsx
+++ b/examples/react/large-file-based/src/routes/__root.tsx
@@ -7,14 +7,16 @@ import {
 import { type QueryClient } from '@tanstack/react-query'
 import { TanStackRouterDevtools } from '@tanstack/router-devtools'
 
-export const Route = createRootRouteWithContext<{ queryClient: QueryClient }>()(
-  {
-    component: RootComponent,
-    notFoundComponent: () => {
-      return <p>Not Found (on root route)</p>
-    },
+export interface Context {
+  queryClient: QueryClient
+}
+
+export const Route = createRootRouteWithContext<Context>()({
+  component: RootComponent,
+  notFoundComponent: () => {
+    return <p>Not Found (on root route)</p>
   },
-)
+})
 
 function RootComponent() {
   return (