--- a/examples/react/large-file-based/src/routes/search/route.tsx
+++ b/examples/react/large-file-based/src/routes/search/route.tsx
@@ -1,6 +1,12 @@
 import * as React from 'react'
 import { createFileRoute } from '@tanstack/react-router'
+import { z } from 'zod'
+
+const search = z.object({
+  rootSearch: z.number(),
+})
 
 export const Route = createFileRoute('/search')({
   component: () => <div>Hello /search!</div>,
+  validateSearch: search,
 })