--- a/examples/react/large-file-based/src/routes/search/searchPlaceholder.tsx
+++ b/examples/react/large-file-based/src/routes/search/searchPlaceholder.tsx
@@ -44,6 +44,7 @@ function SearchComponent() {
           page: 0,
           offset: 10,
           search: 'search',
+          rootSearch: 0,
         }}
       >
         Search